# Reduce spread operators for perf round 2: .parse()