--- a/library/src/schemas/map/mapAsync.ts
+++ b/library/src/schemas/map/mapAsync.ts
@@ -7,8 +7,9 @@ import type {
 } from '../../types.ts';
 import {
   executePipeAsync,
-  getCurrentPath,
   getErrorAndPipe,
+  getPath,
+  getPathInfo,
   getPipeInfo,
 } from '../../utils/index.ts';
 import type { MapInput, MapOutput } from './types.ts';
@@ -123,7 +124,7 @@ export function mapAsync<
       await Promise.all(
         Array.from(input.entries()).map(async ([inputKey, inputValue]) => {
           // Get current path
-          const path = getCurrentPath(info, {
+          const path = getPath(info?.path, {
             schema: 'map',
             input,
             key: inputKey,
@@ -137,7 +138,7 @@ export function mapAsync<
                 // Note: Output key is nested in array, so that also a falsy value
                 // further down can be recognized as valid value
                 return [
-                  await key.parse(inputKey, { ...info, origin: 'key', path }),
+                  await key.parse(inputKey, getPathInfo(info, path, 'key')),
                 ] as const;
 
                 // Throw or fill issues in case of an error
@@ -155,7 +156,7 @@ export function mapAsync<
                 // Note: Output value is nested in array, so that also a falsy value
                 // further down can be recognized as valid value
                 return [
-                  await value.parse(inputValue, { ...info, path }),
+                  await value.parse(inputValue, getPathInfo(info, path)),
                 ] as const;
 
                 // Throw or fill issues in case of an error