--- a/library/src/schemas/object/object.ts
+++ b/library/src/schemas/object/object.ts
@@ -4,6 +4,7 @@ import {
   executePipe,
   getCurrentPath,
   getErrorAndPipe,
+  getParseInfoWithPath,
   getPipeInfo,
 } from '../../utils/index.ts';
 import type { ObjectOutput, ObjectInput } from './types.ts';
@@ -112,10 +113,8 @@ export function object<TObjectShape extends ObjectShape>(
       for (const [key, schema] of Object.entries(object)) {
         try {
           const value = (input as Record<string, unknown>)[key];
-          output[key] = schema.parse(value, {
-            ...info,
-            path: getCurrentPath(info, { schema: 'object', input, key, value }),
-          });
+          output[key] = schema.parse(value, getParseInfoWithPath(info,
+            getCurrentPath(info, { schema: 'object', input, key, value })))
 
           // Throw or fill issues in case of an error
         } catch (error) {