--- a/library/src/schemas/object/objectAsync.ts
+++ b/library/src/schemas/object/objectAsync.ts
@@ -2,8 +2,9 @@ import { type Issue, type Issues, ValiError } from '../../error/index.ts';
 import type { BaseSchema, BaseSchemaAsync, PipeAsync } from '../../types.ts';
 import {
   executePipeAsync,
-  getCurrentPath,
   getErrorAndPipe,
+  getPath,
+  getPathInfo,
   getPipeInfo,
 } from '../../utils/index.ts';
 import type { ObjectInput, ObjectOutput } from './types.ts';
@@ -116,15 +117,18 @@ export function objectAsync<TObjectShape extends ObjectShapeAsync>(
         Object.entries(object).map(async ([key, schema]) => {
           try {
             const value = (input as Record<string, unknown>)[key];
-            output[key] = await schema.parse(value, {
-              ...info,
-              path: getCurrentPath(info, {
-                schema: 'object',
-                input,
-                key,
-                value,
-              }),
-            });
+            output[key] = await schema.parse(
+              value,
+              getPathInfo(
+                info,
+                getPath(info?.path, {
+                  schema: 'object',
+                  input,
+                  key,
+                  value,
+                })
+              )
+            );
 
             // Throw or fill issues in case of an error
           } catch (error) {