--- a/library/src/utils/getCurrentPath/index.ts
+++ b/library/src/utils/getCurrentPath/index.ts
@@ -1,1 +1,0 @@
-export * from './getCurrentPath.ts';