--- a/library/src/utils/getPath/index.ts
+++ b/library/src/utils/getPath/index.ts
@@ -1,0 +1,1 @@
+export * from './getPath.ts';