… library/src/utils/getCurrentPath/getCurrentPath.test.ts
… library/src/utils/getCurrentPath/getCurrentPath.ts