After https://github.com/fabian-hiller/valibot/pull/46, we still have object spread operator when calling `.parse()`. I found this also has negative performance impact.

I'll post benchmarks and will try to update other files as well