@fabian-hiller 2023-08-12
I'll check the changes in 15 minutes and give you feedback. I'll commit a few changes to the main branch that I've made in a few minutes.