@naruaway 2023-08-13
@fabian-hiller, thanks for handling the actual remaining changes 👍 
Now Valibot is faster than Zod for valid data 🎉 
This means probably we should see better score than Zod in https://github.com/moltar/typescript-runtime-type-benchmarks after we publish the new Valibot package since their benchmark seems to be checking only "valid data" case

## vs. Previous Valibot
I ran [naruaway/valibot-benchmarks](https://github.com/naruaway/valibot-benchmarks) using Node.js v20.5.0 and Bun v0.7.2 on my MacBook Air M2 to compare this PR (27278fe99be29dfa96fe950b126ad7fca54218d5) against the main branch (27278fe99be29dfa96fe950b126ad7fca54218d5).

- Significant speed up (more than 100%) is observed for both Node.js and Bun.
- the most notable case is **170% faster for Node.js** :tada: for [`deep` data set](https://github.com/naruaway/valibot-benchmarks/tree/ac99da10597c9b4c98e20a4d1a5df74aba663896/test_data/deep)
- This PR improved performance consistently across the all combinations

<details>
<summary>Valibot in this PR compared against Valibot `main`</summary>
<pre>
Running benchmarks using nodejs
Showing the result when running runtime environment (e.g. Node.js) 100 times for each:
case: "deep", data: "valid"
 170% faster
  baseline: 130783 ops/s, +- 3% stddev
  target: 352828 ops/s, +- 3% stddev
case: "many_features", data: "invalid"
 13% faster
  baseline: 16468 ops/s, +- 2% stddev
  target: 18690 ops/s, +- 2% stddev
case: "many_features", data: "valid"
 38% faster
  baseline: 77907 ops/s, +- 1% stddev
  target: 107824 ops/s, +- 1% stddev
case: "optional_nullable", data: "valid"
 83% faster
  baseline: 207116 ops/s, +- 2% stddev
  target: 378485 ops/s, +- 3% stddev
case: "wide", data: "invalid"
 7% faster
  baseline: 3756 ops/s, +- 4% stddev
  target: 4013 ops/s, +- 4% stddev
case: "wide", data: "valid"
 80% faster
  baseline: 128989 ops/s, +- 4% stddev
  target: 232356 ops/s, +- 4% stddev
Running benchmarks using bun
Showing the result when running runtime environment (e.g. Node.js) 100 times for each:
case: "deep", data: "valid"
 69% faster
  baseline: 230037 ops/s, +- 5% stddev
  target: 388609 ops/s, +- 4% stddev
case: "many_features", data: "invalid"
 7% faster
  baseline: 81143 ops/s, +- 4% stddev
  target: 86581 ops/s, +- 3% stddev
case: "many_features", data: "valid"
 50% faster
  baseline: 264671 ops/s, +- 5% stddev
  target: 397774 ops/s, +- 5% stddev
case: "optional_nullable", data: "valid"
 112% faster
  baseline: 200650 ops/s, +- 7% stddev
  target: 425620 ops/s, +- 9% stddev
case: "wide", data: "invalid"
 10% faster
  baseline: 20165 ops/s, +- 4% stddev
  target: 22141 ops/s, +- 3% stddev
case: "wide", data: "valid"
 95% faster
  baseline: 147967 ops/s, +- 6% stddev
  target: 288962 ops/s, +- 5% stddev
out of 12 cases, 12 cases (100%) were faster
</pre>
</details>



## vs. Zod

I ran a similar benchmark as the above but to compare this PR against Zod

- **:tada: Now Valibot is faster than Zod for any test cases I prepared for valid data**
- Note that when the data is invalid, Valibot is consistently slower than Zod for any engine / schemas

<details>
<summary>Valibot in this PR compared against Zod v3.21.4</summary>
<pre>
Running benchmarks using nodejs
Showing the result when running runtime environment (e.g. Node.js) 100 times for each:
case: "deep", data: "valid"
 110% faster
  baseline: 171819 ops/s, +- 2% stddev
  target: 361375 ops/s, +- 2% stddev
case: "many_features", data: "invalid"
 43% slower
  baseline: 32915 ops/s, +- 3% stddev
  target: 18823 ops/s, +- 2% stddev
case: "many_features", data: "valid"
 21% faster
  baseline: 87467 ops/s, +- 8% stddev
  target: 105528 ops/s, +- 5% stddev
case: "optional_nullable", data: "valid"
 88% faster
  baseline: 201040 ops/s, +- 6% stddev
  target: 378693 ops/s, +- 4% stddev
case: "wide", data: "invalid"
 82% slower
  baseline: 21994 ops/s, +- 6% stddev
  target: 4031 ops/s, +- 6% stddev
case: "wide", data: "valid"
 82% faster
  baseline: 130101 ops/s, +- 8% stddev
  target: 236298 ops/s, +- 5% stddev
Running benchmarks using bun
Showing the result when running runtime environment (e.g. Node.js) 100 times for each:
case: "deep", data: "valid"
 164% faster
  baseline: 148862 ops/s, +- 5% stddev
  target: 392686 ops/s, +- 6% stddev
case: "many_features", data: "invalid"
 57% slower
  baseline: 203677 ops/s, +- 2% stddev
  target: 86844 ops/s, +- 2% stddev
case: "many_features", data: "valid"
 115% faster
  baseline: 192470 ops/s, +- 1% stddev
  target: 413984 ops/s, +- 1% stddev
case: "optional_nullable", data: "valid"
 76% faster
  baseline: 249489 ops/s, +- 2% stddev
  target: 439901 ops/s, +- 6% stddev
case: "wide", data: "invalid"
 70% slower
  baseline: 74880 ops/s, +- 1% stddev
  target: 22697 ops/s, +- 1% stddev
case: "wide", data: "valid"
 57% faster
  baseline: 190875 ops/s, +- 2% stddev
  target: 298863 ops/s, +- 4% stddev
out of 12 cases, 8 cases (67%) were faster
<pre>
</details>